Add files via upload
This file does have every CMPUT class, it's lecture timings (all of them) for both the terms, and it's prerequisites. But I haven't rigorously checked the time conflicts. It SHOULD work but if it doesn't, don't @me. Also, classesList is just a list of all the class objects. Use it if you want. Prerequisites are not parsed (they are just in this format: CMPUTXXXYYYMATHZZZSTATSAAA. Shouldn't be that hard to parse, just search for the substring CMPUT, MATH and STATS and read till EOL.